+++ visualization/app/codeCharta/state/store/fileSettings/attributeTypes/attributeTypes.reducer.spec.ts
@@ -5,8 +5,8 @@ import { AttributeTypes } from "../../../../codeCharta.model"
 
 describe("attributeTypes", () => {
 	const defaultValue: AttributeTypes = {
-		nodes: [],
-		edges: []
+		nodes: {},
+		edges: {}
 	}
 
 	describe("Default State", () => {