+++ CHANGELOG.md
@@ -9,11 +9,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Added 🚀
 
+- Median symbol for aggregated relative metrics #365
+- AttributeTypes for tokeiImporter and SCMLogParser #365
+
 ### Changed
+
+- Metrics with AttributeType relative are now aggregated using the median #365
 
 ### Removed 🗑
 
 ### Fixed 🐞
+
+- Consistency of AttributeTypes representation #365
 
 ### Chore 👨‍💻 👩‍💻